app_dev_backup_20260104_185706/backend/requirements.txt
# FastAPI Backend Requirements
# Data: 2026-01-03

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Data Processing (compatível com Python 3.9)
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
xlrd>=2.0.0
